+++ src/app/app.module.ts
@@ -1,5 +1,4 @@
 import { Module } from '@nestjs/common'
-import { MulterModule } from '@nestjs/platform-express'
 import { CustomersModule } from 'services/customers'
 import { MoviesModule } from 'services/movies'
 import { ShowtimesModule } from 'services/showtimes'
@@ -14,12 +13,10 @@ import {
 } from './controllers'
 import { CustomerJwtStrategy, CustomerLocalStrategy } from './controllers/guards'
 import { CoreModule } from './core'
-import { MulterConfigService } from './multer-config.service'
 
 @Module({
     imports: [
         CoreModule,
-        MulterModule.registerAsync({ useClass: MulterConfigService }),
         CustomersModule,
         StorageFilesModule,
         MoviesModule,
