+++ src/app/services/theaters/models/index.ts
@@ -1,1 +1,2 @@
+export * from './seatmap.model'
 export * from './theater.model'
